Move text_to_entries under the new content folder
