+++ src/khoj/processor/content/github/github_to_entries.py
@@ -11,7 +11,7 @@ from khoj.database.models import Entry as DbEntry
 from khoj.database.models import GithubConfig, KhojUser
 from khoj.processor.content.markdown.markdown_to_entries import MarkdownToEntries
 from khoj.processor.content.org_mode.org_to_entries import OrgToEntries
-from khoj.processor.text_to_entries import TextToEntries
+from khoj.processor.content.text_to_entries import TextToEntries
 
 # Internal Packages
 from khoj.utils.helpers import timer
