+++ src/khoj/processor/content/markdown/markdown_to_entries.py
@@ -10,7 +10,7 @@ from khoj.database.models import Entry as DbEntry
 from khoj.database.models import KhojUser
 
 # Internal Packages
-from khoj.processor.text_to_entries import TextToEntries
+from khoj.processor.content.text_to_entries import TextToEntries
 from khoj.utils.constants import empty_escape_sequences
 from khoj.utils.helpers import timer
 from khoj.utils.rawconfig import Entry
