+++ tests/test_org_to_entries.py
@@ -4,7 +4,7 @@ import os
 
 # Internal Packages
 from khoj.processor.content.org_mode.org_to_entries import OrgToEntries
-from khoj.processor.text_to_entries import TextToEntries
+from khoj.processor.content.text_to_entries import TextToEntries
 from khoj.utils.helpers import is_none_or_empty
 from khoj.utils.rawconfig import Entry
 from khoj.utils.fs_syncer import get_org_files
